backend/pyproject.toml
[project]
name = "agentloom"
version = "0.1.0"
description = "Visual agent workflow platform — every conversation is a branchable DAG"
requires-python = ">=3.12"
authors = [{ name = "usingnamespacestc" }]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "sse-starlette>=2.1.0",
    "python-multipart>=0.0.9",
    "anthropic>=0.40.0",
    "mcp>=1.0.0",
    "PyYAML>=6.0.0",
    "uuid-utils>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.11.0",
    "types-PyYAML",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agentloom*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["../tests/backend"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
ignore_missing_imports = true
